+++ lib/features/password_manager/dashboard/widgets/dashboard_layout/widgets/floating_nav_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../dashboard_layout_constants.dart';
+import '../config/dashboard_layout_constants.dart';
 import 'floating_nav_item.dart';
 
 /// Плавающий навигационный бар со скользящим индикатором